% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-version.R
\name{.url_path}
\alias{.url_path}
\title{Paste components together like file.path, but eliminate consecutive separator chars}
\usage{
.url_path(...)
}
\description{
Paste components together like file.path, but eliminate consecutive separator chars
}
