% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-version.R
\name{package_installed}
\alias{package_installed}
\title{Return whether a package is installed and satisfies criteria}
\usage{
package_installed(pkg, criteria)
}
\arguments{
\item{pkg}{package name}

\item{criteria}{character vector expressing criteria for some version to satisfy.  Options include:
\begin{itemize}
\item \code{NULL} or \code{NA}, indicating that the package must be present, but need not satisfy any
particular version
\item An exact version required, as a string, e.g. \code{"0.1.13"}
\item A comparison operator and a version, e.g. \code{">= 0.1.12"}
\item Several criteria to satisfy, as a comma-separated string, e.g. \code{">= 1.12.0, < 1.14"}
\item Several criteria to satisfy, as elements of a character vector, e.g. \code{c(">= 1.12.0", "< 1.14")}
\end{itemize}}
}
\value{
TRUE if \code{pkg} is already installed, and its version satisfies all criteria \code{criteria}
}
\description{
Return whether a package is installed and satisfies criteria
}
