% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-version.R
\name{version_from_tarball}
\alias{version_from_tarball}
\title{Extract the version from the name of a package tarball}
\usage{
version_from_tarball(tarball_name)
}
\arguments{
\item{tarball_name}{character vector of files or paths from which to extract version numbers}
}
\value{
versions extracted, or \code{NULL} when extraction fails
}
\description{
Extract the version from the name of a package tarball
}
